refactor: core 패키지 변경
