+++ core-bank/src/main/java/com/example/core_bank/core_bank/core/businessnumber/BusinessNumberResponse.java
@@ -1,4 +1,4 @@
-package com.example.core_bank.core_bank.dto.businessnumber;
+package com.example.core_bank.core_bank.core.businessnumber;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
