+++ core-bank/src/main/java/com/example/core_bank/core_bank/core/controller/BusinessNumberController.java
@@ -1,7 +1,7 @@
-package com.example.core_bank.core_bank.controller;
+package com.example.core_bank.core_bank.core.controller;
 
-import com.example.core_bank.core_bank.dto.businessnumber.BusinessNumberResponse;
-import com.example.core_bank.core_bank.service.BusinessNumberService;
+import com.example.core_bank.core_bank.core.businessnumber.BusinessNumberResponse;
+import com.example.core_bank.core_bank.core.service.BusinessNumberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
