+++ core-bank/src/main/java/com/example/core_bank/core_bank/core/model/Bank.java
@@ -1,7 +1,6 @@
-package com.example.core_bank.core_bank.model;
+package com.example.core_bank.core_bank.core.model;
 
 import jakarta.persistence.*;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
