+++ core-bank/src/main/java/com/example/core_bank/core_bank/core/model/TransactionHistory.java
@@ -1,10 +1,8 @@
-package com.example.core_bank.core_bank.model;
+package com.example.core_bank.core_bank.core.model;
 
 import jakarta.persistence.*;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.NonNull;
 
 import java.time.LocalDate;
